+++ src/fold.ts
@@ -8,6 +8,7 @@ import {
 	foldedRanges,
 	unfoldAll,
 	codeFolding,
+	foldService,
 } from '@codemirror/language';
 import {modeConfig} from './config';
 import type {EditorView, Tooltip} from '@codemirror/view';
@@ -328,3 +329,25 @@ export const foldHandler = (view: EditorView) => (e: MouseEvent): void => {
 		dom.remove();
 	}
 };
+
+export const foldOnIndent: Extension = foldService.of(({doc, tabSize}, start, from) => {
+	const {text, number} = doc.lineAt(start);
+	if (!text.trim()) {
+		return null;
+	}
+	const getIndent = (line: string): number => /^\s*/u.exec(line)![0].replace(/\t/gu, ' '.repeat(tabSize)).length;
+	const indent = getIndent(text);
+	let j = number,
+		empty = true;
+	for (; j < doc.lines; j++) {
+		const {text: next} = doc.line(j + 1);
+		if (next.trim()) {
+			empty = false;
+			const nextIndent = getIndent(next);
+			if (indent >= nextIndent) {
+				break;
+			}
+		}
+	}
+	return empty || j === number ? null : {from, to: doc.line(j).to};
+});